feat: separate column controls, add parent page component
